implement & use redis center as notification.center
