+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/optimizely/agent
 
-go 1.20
+go 1.21.0
 
 require (
 	github.com/go-chi/chi/v5 v5.0.8
@@ -12,7 +12,7 @@ require (
 	github.com/golang-jwt/jwt/v4 v4.5.0
 	github.com/google/uuid v1.3.0
 	github.com/lestrrat-go/jwx v0.9.0
-	github.com/optimizely/go-sdk v1.8.4-0.20230515121609-7ffed835c991
+	github.com/optimizely/go-sdk v1.8.4-0.20230905115300-bf10312aeb60
 	github.com/orcaman/concurrent-map v1.0.0
 	github.com/prometheus/client_golang v1.11.0
 	github.com/rakyll/statik v0.1.7
